Intermediate commit 3.
From 1: The vertex doesn’t hold the current config separately, but this
will be the state of the vertex, the signal type is Any. In
Configuration, we introduce two new methods, isSameState and
updateNeighbours. updateNeighbours needs SignalType (because the Rank
versions would update differently), which has to be part of the type
signatures of all classes. The Config type is restricted to be a
subtype of Configuration[Id, Action]. In the target function and the
optimizer we introduce the updateMemory method.